move splitbytes in utils
