+++ src/utils/mod.rs
@@ -1,0 +1,2 @@
+pub mod split_bytes;
+pub use split_bytes::{SplitBytes, SplitBytesFactory};
